backend/requirements.txt
pydantic
fastapi[standard]
langchain
langchain-core
langgraph
uvicorn
python-dotenv
langchain-community
requests
google-search-results
